Revert rect caching and obstructed windows optimizations (#515)
* Revert "Fix the current workspace not being selected during startup (#511)"

This reverts commit 5774eb296d2ed7673a27ff314a111bd846c93a3d.

* Revert "Fix segfault when using cinnamon --replace (#512)"

This reverts commit edb36359e2e5b8afac11f450d7701f01c7d94ee7.

* Revert "window-actor: Use source actor for the close, minimize effects (#508)"

This reverts commit fb2cb1aab2810ad73ddfa1f0ebd1522979835fc8.

* Revert "Fix screensaver regressions (#489)"

This reverts commit c5253b4a05866fc35fc64adfab511e01e2928aa4.

* Revert "window: Fix tile HUD color not changing when holding the snap modifier during a drag (#488)"

This reverts commit a4cff4186e3bb69fa41dfa6eb2718677e00479db.

* Revert "window: Call update_outer_rect in update_resize"

This reverts commit e75e8792ce15a7e8e71516579d6d98b188f1d950.

* Revert "window-actor: Update clip_shadow before the redraw"

This reverts commit 91571814de8f28aec8cb2c342f83e54948321646.

* Revert "Unconditionalize meta_window_update_outer_rect (#484)"

This reverts commit 5bba546a827cf39b1e4a46f9ffa83451ac77f502.

* Revert "window: Fix edge resistance"

This reverts commit ee0af7258bfe28a0fbb495671f60b9b9daf87052.

* Revert "hidpi: Restore scaling to initial cursor size, tile HUD threshold and"

This reverts commit c72054b00847966fdeb975d985557df205dbff38.

* Revert "compositor: Check if priv is NULL before calling meta_window_actor_should_unredirect (#479)"

This reverts commit b625cfbcfb3ce8b11bbaa89a5e8f3850f4046eef.

* Revert " Fix cinnamon-screensaver freezing up when activated (#477)"

This reverts commit d6f5c40871514f2697112ccce8ba0280ec8370c2.

* Revert " window-actor: Support applying effects on clones (#464)"

This reverts commit 75776e74106bea7c81a4b9b7a1422ee4adcab276.

* Revert "window-actor: Add reference to MetaDisplay (#476)"

This reverts commit dae1fe7a31b7e99380efc0895d7aa47476fcb3c9.

* Revert "boxes: Return if the rect pointers are null (#478)"

This reverts commit 4ef006532f061ade3b2cfdf32ac4042c2e8e0c09.

* Revert "Add MetaPrefsState struct for retrieving the prefs state without function invocation (#466)"

This reverts commit 72cf2165259a0ea89da3ec2272fd65b131085d1b.

* Revert "Make sure windows paint before un-fullscreening (#475)"

This reverts commit 355c404e15ab51e36977755eb35430f8c0a387d6.

* Revert "Fix unredirected windows not being clipped out of the stage (#474)"

This reverts commit 1c45ee3279b2fbc9cdbde7ac659239269ff54983.

* Revert "Replace some GLib wrappers (#421)"

This reverts commit 317c06f5a5b3c3f49ad2e55af8cf87e27641bbd2.

* Revert "compositor-private: remove a couple of duplicated lines (#472)"

This reverts commit 8e3c6029880d590b0e2c489f3e374d5874df01d6.

* Revert "Replace some deprecated GDK methods (#443)"

This reverts commit 2a06ea958e100b264d0093f3bfb999a867f30b53.

* Revert "compositor: Fix workspace switching, improve window dragging smoothness (#465)"

This reverts commit 7b97e77b05138b8af8cc8828e3e22312ee3e3b86.

* Revert " Expose the sync method to preferences (#452)"

This reverts commit 40d073df03ae01955e63f3419d71493e1e77738e.

* Revert "window: Remove obsolete cast in meta_window_adjust_opacity"

This reverts commit 3052914e0ad88676ce4736887bd040ff12be3397.

* Revert "window-actor: Use int for opacity"

This reverts commit 8448e692504b5e71dc4d66b36ca68da7afa9a42b.

* Revert "window-actor: Make actor non-reactive (#440)"

This reverts commit 462a5344cb738f09eb8f317b04d36755e4cabaf4.

* Revert "Add meta_display_popup_window_visible (#445)"

This reverts commit 4184eeb91a6358487bd1ff573a4cc34c42a6f43a.

* Revert "Clean up the pre-paint functions (#460)"

This reverts commit 5011f94db0248d8eaed9bbeb4616cb6dfbeca76e.

* Revert "Fix unredirection for wine games, mpv"

This reverts commit 63ddb021f2393610e16f44949e9890fc86e3ad51.

* Revert "Make compositor_private a MetaWindowActor type (#454)"

This reverts commit aeb6395fedda3c0dd94f0a64965cc59182c85104.

* Revert "window-actor: override Clutter opacity property and method (#449)"

This reverts commit 7d4a9da08a81d8e0f98e5612514e3cd8a7402a3b.

* Revert "window-actor: Fix squished window textures (#453)"

This reverts commit f752fb56cdf6b9ab288740e0bbe5a7fd6a682da7.

* Revert "Bump version"

This reverts commit 907f142d3e15e83610a5486afcddd399b7314cab.

* Revert "Revert two commits from #410 (#448)"

This reverts commit b1f346a55fef57028fc9654a9e608688078111ea.

* Revert "compositor: Fix tiling placement (#447)"

This reverts commit 023a6a417192d978f8f9b222c5a70f1a6c4b1e68.

* Revert "window: Update monitor state before showing the window (#444)"

This reverts commit 45498a235f4dd9c7ce8c5cb822541a82d852a5b5.

* Revert "screen: Update the rect cache on monitor change (#441)"

This reverts commit edfc5238c8aca3bd494cada86d358b96134a4d8c.

* Revert "screen: Ensure monitors update on monitor change (#439)"

This reverts commit ad4dbc7712e1e61b83623e2619ec718cee6da734.

* Revert "ui: Don't apply CSD window filtering for button events (#435)"

This reverts commit df879dc3d9c4179654ff2b301a6572ab6e0a3c05.

* Revert "Improve sync between the cursor and frames (#397)"

This reverts commit f9d161e1ef1ac1e2508765d27fee22cf644357a2.

* Revert "Various compositor improvements (#410)"

This reverts commit 48f669aa22c91b0056f912cbc055c7eccff848a4.

* Revert "window: Cache the outer, client rects (#427)"

This reverts commit 009f0f391f9168f48ae7633928347dd539e22035.

+++ src/meta/meta-shadow-factory.h
@@ -49,7 +49,7 @@ struct _MetaShadowParams
   int top_fade;
   int x_offset;
   int y_offset;
-  int opacity;
+  guint8 opacity;
 };
 
 #define META_TYPE_SHADOW_FACTORY            (meta_shadow_factory_get_type ())
